pyproject.toml
[tool.poetry]
name = "qabot"
version = "0.1.0"
description = ""
authors = ["FourZd <shogo.ryaskov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.54.4"
faiss-cpu = "^1.9.0"
streamlit = "^1.40.1"
bs4 = "^0.0.2"
aiohttp = "^3.11.5"
beautifulsoup4 = "^4.12.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
